+++ src/components/form-reference/configured-form-reference.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormReference from './form-reference.component';
 import validate from './form-reference-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'reference',
