+++ src/components/form-reference/form-reference-validations.js
@@ -1,6 +1,6 @@
 import _get from 'lodash/get';
 
-import TranslationsService from '../../../services/translations';
+import TranslationsService from '../../services/translations';
 
 const validate = values => {
   const errors = {};
